Fix ugly sleep() hack in fiji.cxx
Earlier, I did not know how to wait for the Java threads, and therefore
without that ugly sleep(), fiji would quit as soon as ImageJ's main
window was drawn.  The "trick" is to call DestroyJavaVM(), which in
fact does not destroy the virtual machine until all java threads have
terminated.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>