Fiji launcher (MacOSX): Error out without a GUI nor --headless
When running Java on MacOSX, and no GUI is present (as can be detected
by the presence of the environment variable SECURITYSESSIONID), we need
to run in --headless mode.  So error out if the user did not ask for one.

Side note: It is quite unnerving that this case led to a crash earlier,
due to Apple's Java deciding not to find the class ij.ImageJ (probably
the JVM does not even bother to expose the AWT classes without a GUI,
one of which ij.ImageJ extends: java.awt.Frame).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>