Fiji launcher (MacOSX): we must call fiji-tiger
fiji-macosx does not run on Tiger, because it tries to link 64-bit
versions of CoreFoundation and JavaVM, which it does not find.

Instead of falling back to 32-bit, it never executes.

So we now detect if we are running fiji-tiger on Leopard, or fiji-macosx
on Tiger, and in those cases revert to the other program.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>