Fiji launcher: fix special handling of 1st argument ending in '.class'
The Fiji launcher interprets a first argument ending in '.class' as the
desired main class to execute (i.e. to call that class' main() method).

However, due to shortcomings in the implementation, this handlling
kicked in even if the user clearly wanted to call Fake.  This commit
fixes that shortcoming.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>