Windows: show errors longer when falling back to system JVM
Earlier, we used to register an atexit handler that slept 60 seconds so the
user could see the console.  However, if we fall back to system JVM, we
call exec(), where the atexit() handler is not called.

Now, we sleep 5 seconds if the console was opened, but we are going to
exec() a system JVM.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>