Fiji launcher: handle out-of-memory gracefully
We can actually detect an out-of-memory (as opposed to other errors,
such as required-dll-not-found) when creating a java vm.

Unfortunately, it is not possible to try again with a more sensible
heap size, since any second call to JNI_CreateJavaVM() of a process
returns JNI_ERR.

So we resort to exec()ing Fiji itself with a more sensible heap size
argument.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>