Fiji launcher (win64): do not always show the console
The original reason to show the console always on 64-bit Windows
was that the programs compiled without -mconsole seemed to crash.

Now, we know the problem: MinGW64 had a bug whereby the command line
was not accurately passed to main().  The strange thing is that __argc
and __argv _do_ contain the correct values.

Unfortunately, newer MinGW64 seem to be buggy, as the compiled programs
require some libgcc .dll.  So we are stuck with the work-around for now.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>