Fiji (win32): recognize bastardized "precompiled/", too
On Windows, we have to cope with the fact that there are long and short
names.  Just assume that the short name reads "PRECOM~1", which should
be true in all cases (you have to work hard to make this detection fail,
but an alternative solution would be too involved for the benefit).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>