Fiji launcher: fix for Java 1.5 on Linux
On Linux, it appears that the Java Runtime does not properly search its
own library path.  So we have to set LD_LIBRARY_PATH, _and_ we have to
reexec the launcher to make it effective.

Avoid this stunt on Java >= 1.6, indicated by the presence of the
subdirectory jli/ in the lib/<arch>/ directory of the JRE.

Noticed by Albert Cardona in issue 130.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>