Fiji launcher: do not start class path with a path separator
Starting the class path with a path separator inadvertently adds the
current directory to the class path.

As this would mean that classes from the current directory would be
loaded via the system class loader, and as a consequence, no functions
from .jar files in plugins/ and jars/ would be loaded, we cannot allow
that.

Noticed by Larry Lindsey.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>