Do not require super-up-to-date glibc on Linux
There is really nothing in newer glibc that was not in older ones which
the Fiji launcher may depend on.

The two exceptions were the stack protection (now disabled using
-fno-stack-protector) and the IPv6 loopback address (now defined
explicitly).

The remaining symbols that might require a new glibc are forced to the
backwards compatible versions using the .symver trick described here:

	http://www.trevorpounds.com/blog/?p=103

To be able to adjust to future changes, there is a script that can
recreate glibc-compat.h.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>