Fiji launcher: fix mkdir_p()
The earlier logic was broken and failed to create directories as needed,
because the launcher tried to make the directories in the wrong order.
This was hidden by the fact that we usually have files in all the
directories, so parent directories were always created before the child
directories.

But now the file names are constructed incorrectly, having multiple
consecutive slashes, so it showed.

Noticed by quite a number of people due to a script being uploaded to
a new directory.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>