Fiji launcher: strip out redundant Java memory options
When specifying -Xmx=999g -Xmx=23m as Java options, Java does not ignore
the first one, but still fails. So if you managed to specify an
invalid heap size in jvm.cfg (e.g. via Edit>Options>Memory & Threads...)
and try to override it on the command line, it did not work.

With this change, the first option is stripped out, and life is good again.

Noticed by Jacques Pecreaux.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>