Fiji launcher (MacOSX): respect when the user said '32-bit'
When Ctrl+clicking on the Fiji icon, one can select "Get Info" in the
context menu. In the window that pops up, there is a checkbox "Open in
32-bit mode".

When a user checks that checkbox, let's not go ahead and execute the
64-bit Fiji (since including 64-bit architecture in the Fiji launcher
makes it fail on MacOSX 10.4 because no 64-bit Java can be found, we
have to jump through that hoop that we always launch 32-bit first and if
we detect a new-enough MacOSX and 64-bit, we relaunch with 64-bit).

The code to detect that the user wanted to launch in 32-bit more was
adapted from http://blog.timac.org/?p=490.

Gently requested by Erwin Frise (to accomodate for Micro-Manager often
requiring 32-bit mode due to a lack of microscope vendors' support for
64-bit platforms).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>