Fiji launcher: remove the JAVA_HOME override
Trusting JAVA_HOME to be set to a usable JVM has caused many
support headaches due to the common situation where it points
to a JVM without Java3D or with an old or broken version of
Java.  This was further discussed in this thread on the
mailing list:

  http://groups.google.com/group/fiji-devel/browse_thread/thread/29d98d6d1e5bad8/255a0b325c12b413

The one objection to this plan was that for Windows users it
is easier to override which Java is used by an environment
variable than via the --java-home command line argument.  The
plan to alleviate this is that in the future there will be a
menu option to configure the location of the JVM to use, and
this will be stored in jvm.cfg