Fiji launcher: also handle magic file extensions handling via subcommands
For example, when launching 'fiji egads.bsh', we want to launch
Beanshell. However, since we want to make things more configurable
and the launcher leaner (with the goal to make it the IJ2 launcher
in mind), we can put that information also in the config part which
might soon be put into an external file.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>