Fiji launcher: do not require -- to separate out ImageJ1 options
Curtis Rueden pointed out that ImageJ1 has only a couple of options which
we can easily identify and pass to ImageJ1 instead of Java.

So when there is no -- on the command line and we find non-ImageJ1
options, we can now DWIM it.

To make this work, we had to move the automatic -port option.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>