ImageJ Launcher (win): do not try to keep re-attached console open
When we have to open a new console because there was none, we naturally
want the window not to flash but to stay there for a while. But if we did
not even open a new console window but simply attached to an existing one,
we do not want that extra wait (the window will stay open anyway in that
case).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>