ImageJ launcher: avoid NullPointerException
The issue was caused by attempting to System.setProperty with a null
value. We now check for that.

Debugging was a little tricky because the error message was not printed
until DetachCurrentThread was called... so not calling
DetachCurrentThread appeared to fix the issue. But it actually only hid
the error message.