Change some "Fiji" references to "ImageJ"
I did not change them all, because we may not want to change it
unilaterally. Some options may indeed be specific to Fiji, for example.

The main goal for now was to make the application display as "ImageJ"
rather than "Fiji" in Mac OS X's menu bar.

This used to be revision r5216.