ImageJ Launcher (MacOSX): support the 'Open in 32-bit mode' option
Like the Fiji Launcher, we now respect when the user Ctrl-clicked on the
ImageJ icon and selected 'Open in 32-bit mode'.

To prevent a segmentation fault, the ij_dir has to be set before we try
to figure out whether the checkbox was clicked.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r5254.