ImageJ launcher (MacOSX): repeat mmacosx-version-min in linker section
To make sure that our executables are compatible with earlier MacOSX
versions, it is not enough to compile the object files with

	-mmacosx-version-min=10.$MINOR

but the executable has to be linked with that option, too.

In Fiji, the compilation/linkage was performed in one step, but with the
NAR plugin the compilation and linking are separate, with distinct
option sets. So we need to repeat the relevant options.

Noticed by Kevin Eliceiri.

Diagnosed with the help of Melissa Linkert.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r5273.