ImageJ launcher: handle update/ first thing after discovering ijdir
Discovered by Jan Eglinger and diagnosed with the help of Melissa
Linkert: the recent update to use the ImageJ launcher requiring
jars/ij-launcher.jar of course demands that the file is moved over from
update/ already before discovering jars/ij-launcher.jar (or whatever
versioned file is available).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

This used to be revision r5438.