Make ClassLoaderPlus interoperable with Retrotranslator
When ClassLauncher & ClassLoaderPlus are loaded by Retrotranslator's
just-in-time-translating class loader, we should make it possible to
continue to use that class loader.

But the earlier assumption was that we can rely on the class loader
responsible for loading all the classes from $IJDIR/jars/ and
$IJDIR/plugins/ being a ClassLoaderPlus.

The only reason we needed that, really, was because URLClassLoader's
addURL() method is protected.

Now, Retrotranslator's TranslatingJarClassLoader is a URLClassLoader, but
not a ClassLoaderPlus.

So as an ugly work-around, teach ClassLoaderPlus to work with
URLClassLoaders, too. In case the one at hand is a ClassLoaderPlus, just
use the now-public addURL() method. Otherwise, fall back to accessing the
method via reflection, after calling setAccessible(true).

That way, we can still run in restricted environments where reflection is
not an option, but support the poor puppies stuck with Java5, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>