Launcher: Auto-activate Retrotranslator for Java < 1.6
We already did that on MacOSX, because we did not ship Java for MacOSX
with Fiji. But Windows and Linux were different; with ImageJ2, we will
not ship any Java version but reuse the system-wide installation. Since
we now have code to discover the system Java's location, we can also
determine its version very easily.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>