ImageJ launcher: handle versioned file names gracefully
We want to use the ImageJ updater and launcher in Fiji, and to be able
to use, say, javassist-1.16.1-GA.jar for the IJ1 patcher, the launcher
needs to wisen up a little when it comes to versioned file names.

This code is based on the code in the ImageJ updater and at least
the version pattern needs to stay synchronized.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>