Launcher: let local jre/ have precendence over all other options
The jre/ subdirectory was established as the place where the JRE lives
by the ImageJ 1.x installer.

Obviously, this cannot support multi-platform installations (e.g. a
portable application for all supported platforms).

Hence we need to make sure that the jre/ subdirectory actually contains
a valid libjvm before going with it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>