Launcher (MacOSX): really strip out the -psn_* command-line options
On MacOSX, the Finder appends a -psn_* command-line parameter to
everything that was double-clicked. This tries to pass on the
information about the the Carbon Process Manager's idea what the current
process is (apparently needed for things such as keeping track of Dock
icons) which has only a loose relation to real BSD processes (according
to 'Mac OS X Internals: A Systems Approach': 'Each CPM process maps to a
BSD process, but not vice versa'). For the record, 'psn' stands for
Process Serial Number. Fun, fun.

Now, already back when the launcher was still exclusive to Fiji, we
realized that it was not a good idea to pass that -psn option to the
Java Virtual Machine because it does not understand that option, so Greg
Jefferis provided a patch to strip it out.

It is an even worse idea to pass the -psn option to the 'java'
executable -- which does understand this option but correctly sees a
discrepancy between the BSD process ID and the Process Serial Number and
fails.

Now, the stripping out was a little bit faulty. It did not quite look
for the last option to be that -psn_* option, but for the *first*. That
fails e.g. when we tried to re-exec due to too much reserved heap space
to accomodate for.

This was reported by Curtis Rueden here:

	http://fiji.sc/bugzilla/show_bug.cgi?id=532

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>