Stop hard-coding fiji-compat dependencies
This changes the way the class path is built: instead of listing the
dependencies explicitly, let's inspect the .jar's manifest. Since even
MiniMaven learnt about adding the Class-Path entry to the manifest,
we can pretty much rely on that entry being present.

The only thing we need to be careful about: as ImageJ2 and Fiji
ship with the version encoded into the .jar files' file names, it is
possible that the version of the shipped file changed, and hence the
default JRE mechanism would no longer add the appropriate class path
elements automatically.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>