Do not patch ImageJ 1.x when running the updater
The command-line updater is the last line of defense when trying to
fix a Fiji installation in case the legacy patching goes wrong. In
that case, it is not a good idea to patch ImageJ 1.x when running
the command-line updater. It is not necessary, anyway, so simply do
not do it.

Noticed by Aryeh Weiss.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>