Fail if java path search finds an Apple framework
One of the ways the launcher attempts to locate a JRE is to look for
where in the PATH the "java" command comes from, including tracing back
through symlinks. Unfortunately, on OS X, this often ends up in the
Java Framework's Commands folder rather than in a proper JRE structure.