Support more than one class path element in legacy mode
As reported by Joachim Wesner, there can be multiple entries in an
ImageJ.cfg's -cp parameter. His example was an ImageJ.cfg looking like
this:

	.
	jre\bin\javaw.exe
	-Xmx338m -cp ij.jar;jython.jar;idl ij.ImageJ

Therefore, just split the parameter by semicolons and pass the parameters
as individual class path elements.

We still have to split them and process them individually because -- in
contrast to the old launcher -- we do not enforce changing the current
working directory to the launcher directory.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>