Allow compiling the ImageJ launcher on MacOSX w/o Java 6
Up until Java 6, Java on MacOSX was maintained (for some creative
definition of the word) by Apple, in the form of a framework. These
frameworks are similar to libraries in the same way as MacOSX .app
directories are similar to applications: slightly broader concepts,
just incompatible enough to offer a nice vendor lock-in.

Starting with Java 7, however, Oracle was tasked with maintaining
Java for MacOSX, like they (and before them, Sun Microsystems) did
with all normal operating systems. Java 7 is therefore not a MacOSX
framework anymore, therefore we have to use the same method of
calling Java as everywhere else: dlopen().

To maintain backwards-compatibility, we need yet another MacOSX
specific profile, which we call MacOSX-Lion (thank you, thank you,
I find it rather inventive a term myself).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>