Process the ij1.plugin.dirs property
Fiji introduced the feature where $HOME/.plugins/ is an additional,
per-user plugin directory. This can be overridden by setting the
'ij1.plugin.dirs' system property to a colon-separated list of paths (on
Windows, semi-colon-separated, because Windows has to do everything just
a little different, just a little) to look for plugins.

Since we will switch ImageJ2 to start up in legacy mode -- which, for
technical reasons, requires the same class path to be used for the
plugins as is used for the ImageJ 1.x core classes -- we need to support
this in the ImageJ launcher, too, so that the class path can be set
appropriately.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>