Avoid using the JavaVM framework on MacOSX
The problem is that newer Java versions are no longer shipped as
frameworks, and consequently the ImageJ launcher triggers an installation
of Java 6 – even on machines that want Java 7 or even Java 8.

So let's violate the obsolete suggestion to use the JavaVM framework, and
access libjvm just like everywhere else: using dlopen().

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>