Move extern declarations to common.h
Proper extern use is to declare an extern variable once in a header
file, define the variable once in a source file, and then #include the
declaring header in each additional source file that would like to
consume the external variable.

We now comply with this use pattern, which fixes an issue where each
source file was retaining its own copy of extern variables (e.g. debug,
info) thus causing debugging info not to be printed.

Source:
http://stackoverflow.com/questions/1433204/
how-do-i-use-extern-to-share-variables-between-source-files-in-c