Add nar.arch.path for ant copy
It is possible for multiple architectures to be built in the same
directory. This causes "**/imagej-launcher" to not be specific enough to
match a single file, which breaks the ant copy task.

Assuming each build directory is distinct per OS, the simplest way to
differentiate the executables is by architecture. Thus a nar
architecture property is now defined by each platform-specific profile,
and then consumed by the ant copy path.

If this proves insufficient we could move to full AOL but that may be
harder to determine in the pom.