Decrease WIN32 max heap size to 1224m
I have tested for the maximum heap size on two different machines. A
Windows 10 (32-bit) VM still works with 1408m, while the launcher crashes
on a Windows 10 (32-bit) physical machine. On the latter machine, I have
determined the max heap size to be 1224m. This launcher also works
properly on the former VM.