Remove Java extension path logic
The java.ext.dirs property is no longer supported in Java 9, and
in fact the JVM refuses to start up if this property is set.
Rather than add case logic for Java 9 vs. earlier versions, this
commit removes all related logic.

This means that support for Java 6 will not fully work anymore
(e.g., the Java-6 version of Java 3D will no longer be available
to the JVM while installed into lib/ext). Therefore, builds of
the ImageJ launcher from this point forward should not be used
with legacy ImageJ installations still running with Java 6.

But it also means that each JVM's default lib/ext paths (if any)
will always kick in, so some related errors will also be
magically solved, such as a missing or mismatched JavaScript
extension when multiple JVM versions are installed -- the ImageJ
launcher has/had a bug here that caused the java.ext.dirs to
point to the lib/ext folder of a non-active JVM in some cases.