Replace native splash with Java implementation
This removes the native implementation that was hard to maintain in the
past and wasn't working properly in some cases. In addition, changes to
how libsplashscreen is linked starting with Java 9 has made using libjvm
basically impossible.

The new Java implementation offers a progress bar during startup which
is currently static. Future versions could improve on this and make the
startup even more verbose s.t. we can get more information on where it
grinds to a halt / an issue appears.