Add guess_java_version(path)
This includes a change to circumvent issues where the version of a
system JDK and the bundled JDK does not match. Since the paths to
libjvm depend on the Java version, we have to do parameterize the
guessing of the Java version with the path for which to guess. This
allows for an explicit guessing of the bundled Java first and the
system Java installations second.