Get rid of guess_java_version
We should scan for the JVM library, not guess at the Java version from
its name. Guessing from the directory name is untenable because some
JVMs simply do not have enough information in the name itself. The user
might even have renamed it to something simple like just "java".

This adds a new requirement: bundled Javas MUST be in a
platform-specific subdirectory of the /java folder, following a naming
conventtion of "win32", "win64", "macosx", "linux64" or "linux32".

Shipping multiple platforms in a single bundle is still OK, as only the
platform-appropriate directory will be searched.

We now search for all known permutations of the java shared library, in
hopes of facilitating universal support on the big 5 OS's.

Co-authored-by: Mark Hiner <hinerm@gmail.com>