Always use ANSI LoadLibrary call
C system functions often have an "A" and "W" versions. "A" indicates
ANSI with fixed 1-byte length characters, while "W" is for Unicode
characters and can be 2-4 bytes. Without specifying, "LoadLibrary"
defaults to Unicode... typically.

On Windows 10 Home, "LoadLibrary" seems to default to the ANSI version,
while on Windows 10 Pro it goes to Unicode. The jvm.dll of the JRE
supplied with Fiji for many years can not be loaded via "LoadLibraryW".
This created a scenario where the launcher appeared to work on W10Home
but not Pro:
https://forum.image.sc/t/new-fiji-launcher-console-window-in-win64/44349

We can resolve this by always loading with the ANSI method, which also
appears to work with more modern JREs.