check-java-version: improve subscript behavior
When check-java-version is called from another script
(run-java-gauntlet) it runs in a subshell, which collects std
output/error from imagej but not from Java. Thus the actual output of
the WhichJava program is lost to the run-java-gauntlet program and thus
cannot be used to verify the pass/fail status of the test.

By instead writing the output to a temporary file and catting the
result, the check-java-version script essentially "produces" the Java
program's output to stdout, while maintaining stderr from the native
launcher (which is desired to preserve debug output from the launcher).

As the stderr and stdout are now both present, note that the order of
WhichJava output has been adjusted so that the actual Java used appears
on the last line (to simplify pass/fail testing).