file-funcs: use access rval
access returns -1 (true) when the file is not found and 0 (false) when
the file exists.

So while boolean negation is technically accurate it is also confusing
to read.

Separately validating the rval also allows for interrogation of the
ERRNO if desired.

See https://pubs.opengroup.org/onlinepubs/009695399/functions/access.html